Implement Fake Pthreads
This commit adds an implementation for "fake" pthreads which behave
like a pthread but are not in the same thread group as the parent
(so that they can be used to ptrace).

The implementation works by first creating a real pthread then a new
non-pthread (spawned using the clone system call) then having the real
pthread save all of its registers and exit abruptly (so that no cleanup
is performed). The non-pthread then loads all of the register values
in order to 'jump' to the position where the real pthread exited from.

The reason for performing this complex sequence is to get glibc to
perform setup of the thread local storage space which each pthread must
have. A portion of the thread local storage is completley under control
of glibc and glibc takes full advantage of this fact in order to store
a wide variety of data structures that are necessarry to the
funcptioning of libc. Since user code normally has no way to access
this space glibc is free to use the space however it wishes and is not
requuired to setup the space consistently across versions of glibc.
Setup of the space is performed during pthread creation and so the
only way to setup the space correctly (outside of pain-stakingly
reimplementing glib's behaviour for every supported version of
glibc) is to get glibc to setup a thread local storage space for us
then hand that space over to our non-pthread thread.

Note that even this implementation is not certain to work as intended.
I have yet to encounter any issues in testing so far but there may
still be some serious flaws with this approach which are yet to be
discovered. In particular, if there are any inconsistencies between the
kernel's view of the fake pthread and the data stored in the thread
local storage space (and glibc is able to discover the inconsistency
by accessing kernel interfaces) then problems may arise.