Fix thread-jump implementation
The critical parts of thread-jump are now all implemented inside
of the assembly code. With this change, the code is now able to
(and does) ensure that, after setting the flag in the jump buffer
to indicate that it is okay to jump, the caller no longer
access the stack which it formerly owned.

A consequence of this is that the tj_set function is no longer
available because it is mandatory for the caller to *immediatley*
vacate  it's stack after writing the jump buffer

- jumping directly to it's target in the case of tj_swap
- exiting immediatley in the case of tj_set_and_exit