Change pthread state to joinable
Change the pthread state to joinable. According to the manpage,
with the detached attribute, resources will be cleaned up when the
thread exits.
It doesn't seem that this is possible if the exit system call is
called directly, nonetheless, it seems that there is no point to
making the thread non-joinable.