Fake the value of getpid()
Overrode the getpid glibc function in order to fool any program which
might save the pid of a forked or cloned child process as returned by
the call to fork or clone and later compare it to the pid returned by
getpid(). The py-trace program thread swaps the main thread of every
process that it attaches to with that of a thread which it spawns and
so any program which performs the aforementioned comparison can detect
an anomaly.

This exact issue was causing firefox's tab processes to crash.

Note that this issue still exists for programs which might call the
getpid system call directly instead of by going through glibc. Once
syscall modification is working this fix needs to be applied at the
syscall level.