Add no_thread debug mode
Added the no_thread build target which builds a variant of py-trace
which simply forks and execs the target process and does not setup
a thread.

This mode is useful for testing ptrace issues in isolation.

If py-trace eventually acquires features which rely on the thread
model, then this build target may have to be removed.