Print some syscalls in human readable format
Implemented printing of some syscalls in a human readable format.

At exit of the syscall (note that now since nothing is printed at
syscall enter, syscalls which never exit will not be displayed)
we now print out a human readable representation of the syscall for
some specific calls.

This has been done to demonstrate that we can view the system calls,
and the arguments to the system calls and also that we can directly
read the memory of traced process (note that currently we will actually
segfault if we attempt to read memory that we are not allowed to access
for any reason and that if the target program uses a clone call to
implement forking we will attach to the child process but will not
be able to read its memory).