Compile seperate so for PRELOAD
Modern gcc does not support linking of an executable so. Therefore we
now need a seperate executable and so. One to read options and setup for
tracing the target, and another to load into the target executable.

This commit, only creates the so and compiles without errors. The
loading of the so into the target is not yet working.