Implement ghost stdio to avoid conflicts with libc
Implement a simplified stdio which can be used in place of glibc stdio
in order to avoid conflicts with glibc stdio. In some cases, glibc stdio
functions perform malloc operations which in turn acquire global locks.
And so when we ptrace the target we interrupt the execution of the stdio
function while the lock is held.

This situation seems to happen when writing to FILE* which point to a
file on disk, but not when FILE* points to a TTY (something to do with
the default buffering settings, probably).

To workaround this, ghost stdio implements some of the more useful stdio
functions and uses the ghost heap for buffering so as to avoid conflicts
with glibc. With this fix, we can now strace the target while
redirecting stderr to a file, were before if we were to try this it
would result in a deadlock.

As of this commit, printf and (most useful) variants are mostly
implemented but formattin of floating point numbers is TBD.