Callback to lua on ptrace events
Also add config to setup the lua lsp. Although somewhat editor specific,
these are checked into source control because any user of this program
is likley to want these files to develop their own lua code.

Without lua definition files, a lua IDE is likley to flag the lua
scripts for use of undefined code, because much of the lua functions
and data avaiable in ghost script is defined in C where the lua IDE is
unable to see them.