Replace some libc syscalls with custom ones
Some libc syscalls cause problems when run in certain targets. In
chromium, for instance, calls to malloc and isatty cause segfaults when
run.
It seems that this has something to do with the PLT. Normally this
should be setup by the dynamic loader, not libc, and so I don't
understand what the problem is here. It wouldn't be entirely suprising
that something like chromium is doing something weird with its dynamic
symbol loading, however.